fix an infinite loop when the token dispenser is having issues
When searching or downloading apks, gplaycli will happily keeping to
get a token over and over from the token dispenser. If the token
dispenser spits out tokens but google doesn't like them, it will
create an infinite loop.

This breaks out of that loop after 5 attempts, and also suggest to
the user that they try using their own credentials. I'd like to
point them to a manpage or README with instructions, but we're
not there yet.